・戻るボタン押下時にダイアログ表示が出る不具合の修正 ・萬田さん提供の画像ファイルの選択表示機能（一覧以外）、及びDBへの登録機能追加

+++ app/src/main/java/com/example/esdgspro/EditProductActivity.kt
@@ -1,15 +1,20 @@
 package com.example.esdgspro
 
-import android.app.Activity
 import android.app.DatePickerDialog
+import android.content.ContentValues
 import android.content.Intent
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.graphics.drawable.BitmapDrawable
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.*
 import androidx.appcompat.app.AlertDialog
+import java.io.ByteArrayOutputStream
 import java.time.LocalDateTime
 
 class EditProductActivity : AppCompatActivity() {
@@ -37,10 +42,18 @@ class EditProductActivity : AppCompatActivity() {
 
         barcodeId.text = id.toString()
 
+        var selectProImg = SdgsCommonLogic()
+
+
         //DB接続用
         val helper = DBHelper(this@EditProductActivity)
         val db = helper.writableDatabase
-        val itemData = db.rawQuery("select * from food_ingredient_tb where ingredient_id = '" + id + "';", null)
+        val itemData = db.rawQuery(
+            "select ingredient_id,ingredient_name," +
+                    "product_class,purchase_date," +
+                    "expiry_date,quantity,state,image " +
+                "from food_ingredient_tb " +
+                    "where ingredient_id = '" + id + "';", null)
 
         image1 = findViewById(R.id.imageView)
         textView = findViewById(R.id.productName)
@@ -66,7 +79,11 @@ class EditProductActivity : AppCompatActivity() {
             while (it.moveToNext()){
                 with(it){
                     if (getBlob(7) != null) {
-                        image1.setImageResource(getInt(7))
+
+                        val blob: ByteArray = getBlob(7)
+                        val bitmap = BitmapFactory.decodeByteArray(blob, 0, blob.size)
+
+                        image1.setImageBitmap(bitmap)
                     }
                     else{
                         image1.setImageResource(R.drawable.camel)
@@ -78,6 +95,24 @@ class EditProductActivity : AppCompatActivity() {
                     expiryDateText.text = getString(4)
                     quantity.text = getInt(5).toString()
                 }
+            }
+        }
+
+        /*「商品分類」選択時*/
+        spinner.onItemSelectedListener = object:AdapterView.OnItemSelectedListener{
+            override fun onItemSelected(
+                parent: AdapterView<*>?,
+                view: View?,
+                position: Int,
+                id: Long
+            ){
+                var curSpnVal = (parent as Spinner).selectedItemPosition + 1
+
+                //selectProImg.curSpn = curSpnVal
+                image1.setImageResource(selectProImg.selectProductImage(curSpnVal))
+            }
+            override fun onNothingSelected(parent: AdapterView<*>?) {
+                //
             }
         }
 
@@ -109,42 +144,45 @@ class EditProductActivity : AppCompatActivity() {
 
     //オプションメニュー：削除ボタン　クリック時
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        AlertDialog.Builder(this)
-            .setTitle(R.string.dialog_title)
-            .setMessage(R.string.dialog_del_message)
-            .setPositiveButton(R.string.dialog_ok) { dialog, which ->
-                val helper = DBHelper(this@EditProductActivity)
-                val db = helper.writableDatabase
-                val id = barcodeId.text
-                val sqlDelete = "delete from food_ingredient_tb  where ingredient_id = '${id}'"
-                val deleteData = db.compileStatement(sqlDelete)
-                when(item.itemId){
-                    R.id.action_delete ->
-                    deleteData.executeUpdateDelete()
+        if (item.itemId == android.R.id.home) {
+            finish()
+        }
+        if(item.itemId == R.id.action_delete){
+            AlertDialog.Builder(this)
+                .setTitle(R.string.dialog_title)
+                .setMessage(R.string.dialog_del_message)
+                .setPositiveButton(R.string.dialog_ok) { dialog, which ->
+                    //ダイアログでOKをクリックした場合、削除を行う。
+                    val helper = DBHelper(this@EditProductActivity)
+                    val db = helper.writableDatabase
+                    val id = barcodeId.text
+                    val sqlDelete = "delete from food_ingredient_tb  where ingredient_id = '${id}'"
+                    val deleteData = db.compileStatement(sqlDelete)
+                    when(item.itemId){
+                        R.id.action_delete ->
+                            deleteData.executeUpdateDelete()
+                    }
+                    finish()
+
+                    //削除完了メッセージ表示
+                    val toast = Toast.makeText(this@EditProductActivity,R.string.cmp_del_message,Toast.LENGTH_LONG)
+                    toast.show()
+
+                    //メニューに戻る
+                    val intent: Intent = Intent(this@EditProductActivity,
+                        MainActivity::class.java)
+                    startActivity(intent)
+
                 }
-                finish()
+                // Cancelの時は何もしない
+                .setNegativeButton(R.string.dialog_cancel) { dialog, which ->
 
-                //メニューに戻る
-                val intent: Intent = Intent(this@EditProductActivity,
-                    MainActivity::class.java)
-                startActivity(intent)
+                }
+                .show()
+        }
 
-            }
-
-            // Cancelの時は何もしない
-            .setNegativeButton(R.string.dialog_cancel) { dialog, which ->
-
-            }
-            .show()
         return true
     }
-        /*override fun onOptionsItemSelected(item: MenuItem): Boolean {
-            println("DELETE")
-            val dialog = SampleDialogFragment()
-            println(dialog)
-            dialog.show(supportFragmentManager, "simple")
-            return super.onOptionsItemSelected(item)
-        }*/
 
 
     private fun showDatePicker(getText: String, getTextValue: TextView) {
@@ -197,7 +235,7 @@ class EditProductActivity : AppCompatActivity() {
     fun updateData(view: View) {
 
         val status: Int
-        val user: String = "shimizu"
+        val user: String = "Update"
 
         if (quantity.text.toString().toInt() > 0){
             status = 0
@@ -210,21 +248,58 @@ class EditProductActivity : AppCompatActivity() {
             .setTitle(R.string.dialog_title)
             .setMessage(R.string.dialog_upd_message)
             .setPositiveButton(R.string.dialog_ok) { dialog, which ->
-                //OKの場合、更新を行う。
+                //ダイアログでOKをクリックした場合、更新を行う。
                 val helper = DBHelper(this@EditProductActivity)
                 val db = helper.writableDatabase
 
-
+/*
                 val dataArray = arrayOf(textView.text, spinner.selectedItemPosition + 1, purchaseDateText.text, expiryDateText.text, quantity.text.toString().toInt(), status, user, barcodeId.text)
                 val sqlUpdate = "update food_ingredient_tb set ingredient_name = '${dataArray[0]}', product_class = ${dataArray[1]}, purchase_date = '${dataArray[2]}', expiry_date = '${dataArray[3]}', quantity = ${dataArray[4]}, state = ${dataArray[5]}, upd_date = CURRENT_DATE, upd_user = '${dataArray[6]}' where ingredient_id = '${dataArray[7]}'"
                 val updateData = db.compileStatement(sqlUpdate)
                 updateData.executeUpdateDelete()
                 finish()
+*/
+                /*画像ファイルをBlobに変換*/
+                val bitmapDrawable = image1.drawable as BitmapDrawable
+                val bitmapDrawablex = bitmapDrawable.bitmap
+                val byteArrayOutputStream = ByteArrayOutputStream();
+
+                bitmapDrawablex.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream)
+                val bytes = byteArrayOutputStream.toByteArray()
+
+                try {
+                    /*登録値を設定*/
+                    val values = ContentValues()
+                    values.put("ingredient_name", textView.getText().toString())
+                    values.put("product_class", spinner.selectedItemPosition + 1)
+                    values.put("purchase_date", purchaseDateText.getText().toString())
+                    values.put("expiry_date", expiryDateText.getText().toString())
+                    values.put("quantity", quantity.getText().toString().toInt())
+                    values.put("state", status)
+                    values.put("image", bytes)
+                    values.put("upd_user", user)
+                    values.put("upd_date",LocalDateTime.now().toString())
+
+                    val whereClauses = "ingredient_id = ?"
+                    val whereArgs = arrayOf(barcodeId.text.toString())
+
+                    /*更新処理実行*/
+                    db.update("food_ingredient_tb", values, whereClauses,whereArgs)
+                }catch(exception: Exception) {
+                    Log.e("updateData", exception.toString())
+                }
+
+                finish()
+
+                //更新完了メッセージ表示
+                val toast = Toast.makeText(this@EditProductActivity,R.string.cmp_upd_message,Toast.LENGTH_LONG)
+                toast.show()
 
                 //メニューに戻る
                 val intent: Intent = Intent(this@EditProductActivity,
                     MainActivity::class.java)
                 startActivity(intent)
+
 
             }
 
@@ -239,7 +314,7 @@ class EditProductActivity : AppCompatActivity() {
 
 
 
-
+/*
     class SpinnerActivity : Activity(), AdapterView.OnItemSelectedListener {
 
         override fun onItemSelected(parent: AdapterView<*>, view: View?, pos: Int, id: Long) {
@@ -251,4 +326,6 @@ class EditProductActivity : AppCompatActivity() {
             // Another interface callback
         }
     }
+
+ */
 }
